+++ src/components/Layout/SharedLayout/SharedLayout.jsx
@@ -4,16 +4,21 @@ import { Header } from 'components/Layout/Header/Header';
 import { Footer } from 'components/Layout/Footer/Footer';
 
 import Loader from '../../../ui/Loader';
+import { StyledBackground } from './SharedLayout.styled';
+import { useSelector } from 'react-redux';
 
 const SharedLayout = () => {
+  const authenticated = useSelector((state) => state.auth.authenticated);
   return (
-    <>
-      <Header />
-      <Suspense fallback={<Loader />}>
-        <Outlet />
-      </Suspense>
-      <Footer />
-    </>
+    <StyledBackground>
+      <div className={authenticated ? 'mainBackground' : 'welcomeBackground'}>
+        <Header />
+        <Suspense fallback={<Loader />}>
+          <Outlet />
+        </Suspense>
+        <Footer />
+      </div>
+    </StyledBackground>
   );
 };
