+++ src/redux/auth/auth.reducer.js
@@ -1,13 +1,5 @@
-import { createAsyncThunk, createSlice, isAnyOf } from '@reduxjs/toolkit';
-import axios from 'axios';
-
-export const instance = axios.create({
-  baseURL: 'https://drink-master-api-umd0.onrender.com/api',
-});
-
-const setToken = (token) => {
-  instance.defaults.headers.common.Authorization = `Bearer ${token}`;
-};
+import { createSlice, isAnyOf } from '@reduxjs/toolkit';
+import { loginThunk, registerThunk } from './auth.operations';
 
 const initialState = {
   userData: null,
@@ -16,68 +8,6 @@ const initialState = {
   authenticated: false,
   token: null,
 };
-
-export const registerThunk = createAsyncThunk(
-  'auth/register',
-  async (formData, thunkApi) => {
-    try {
-      const { data } = await instance.post('/auth/signup', formData);
-      console.log('data: ', data);
-      setToken(data.token);
-      return data;
-    } catch (err) {
-      return thunkApi.rejectWithValue(err.message);
-    }
-  }
-);
-
-export const loginThunk = createAsyncThunk(
-  'auth/login',
-  async (formData, thunkApi) => {
-    try {
-      const { data } = await instance.post('/auth/signin', formData);
-      setToken(data.token);
-      return data;
-    } catch (err) {
-      return thunkApi.rejectWithValue(err.message);
-    }
-  }
-);
-
-// export const logOutThunk = createAsyncThunk(
-//   'auth/logOut',
-//   async (_, thunkApi) => {
-//     try {
-//       const { data } = await instance.post('/auth/signout');
-//       return data;
-//     } catch (err) {
-//       return thunkApi.rejectWithValue(err.message);
-//     }
-//   }
-// );
-
-// export const refreshThunk = createAsyncThunk(
-//   'auth/refresh',
-//   async (_, thunkApi) => {
-//     try {
-//       const state = thunkApi.getState();
-//       const token = state.auth.token;
-//       setToken(token);
-//       const { data } = await instance.get('/users/current');
-//       return data;
-//     } catch (err) {
-//       return thunkApi.rejectWithValue(err.message);
-//     }
-//   },
-//   {
-//     condition: (_, thunkApi) => {
-//       const state = thunkApi.getState();
-//       const token = state.auth.token;
-//       if (!token) return false;
-//       return true;
-//     },
-//   }
-// );
 
 const authSlice = createSlice({
   name: 'auth',
